Merge pull request #31 from TeoZosa/ensure-user-install-steps-correspond-to-package-under-test
🐛 Ensure CI PyPI Package Installations Correspond to Package-Under-Test

The previous setup had a race condition between the `verify-user-install` CI workflow job installing the latest PyPI package and the `release` Release workflow job uploading the package corresponding to the current commit. This PR synchronizes those jobs, making `verify-user-install` depend on the Release workflow finishing successfully, and adding a `Validate published package is reachable` step to the Release workflow `release` job to ensure that the Release workflow does not finish until the uploaded package can be successfully retrieved on the Python package indexes (note: given the versioning logic, versions are guaranteed to be ascending and, therefore, the latest package will correspond to the package built in the associated CI/CD run).

Original-Commit: TeoZosa/cookiecutter-cruft-poetry-tox-pre-commit-ci-cd@e13493f