:recycle: Tightly couple env var with sole corresponding target (`test`)
Since `TOX_PARALLEL_NO_SPINNER` is only relevant to tox invocations that
 use the `-p`/`--parallel`, removing this environment variable from the
global scope and tightly coupling it to the `test` target (the only
dependent target) explicitly communicates that this is the only place
where this variable is used. In other words, it removes any confusion
around if `TOX_PARALLEL_NO_SPINNER` affects the behavior of any other
targets.

Signed-off-by: Teo Zosa <teofilo@sonosim.com>

Original-Commit: TeoZosa/cookiecutter-cruft-poetry-tox-pre-commit-ci-cd@3c63c3d