:cookie: :memo: Document Make's role in project
Interface shims leverage the
[dependency inversion principle](https://en.wikipedia.org/wiki/Dependency_inversion_principle)
to allow for implementation detail flexibility, interface extensibility
and, in this case, a uniform interface across projects.

Ex. from "The Missing README"
- Chapter 11: Creating Evolvable Architectures
- Section: Design for Evolvability
- Sub-section: You Ain’t Gonna Need It (para. 9)

"In the meantime, place interface shims where you suspect optimizations
can be inserted, but don’t actually implement them. For example, if you
are creating a new file format and suspect you’ll need compression or
encryption later, provide a header that specifies the encoding, but only
 implement the uncompressed encoding." [1]

[1] C. Riccomini and D. Ryaboy, The Missing README: A Guide for the New Software Engineer, Paperback. No Starch Press, 2021.

Original-Commit: TeoZosa/cookiecutter-cruft-poetry-tox-pre-commit-ci-cd@b83ef09